Backend/requirements-base.txt
# FastAPI and Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.8.0
pydantic-settings>=2.1.0
asyncpg>=0.29.0

# Database Drivers
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
motor>=3.3.2
pymongo>=4.6.0
redis>=5.0.1
weaviate-client>=4.4.0
neo4j>=5.15.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Resume Parsing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfplumber>=0.10.3

# Utilities
python-dotenv>=1.0.0
